\name{ELISA}
\alias{ELISA}
\docType{data}
\title{Enzyme-Linked ImmunoSortbent Assay Data (ELISA)}
\description{
A data set containing with the results from an ELISA experiment
}
\usage{data(ELISA)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Concentration}}{Concentration analyte}
    \item{\code{Response}}{Response of assay}
  }
}

\examples{
data(ELISA)
}
\keyword{datasets}
