\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,ANY,ANY-method}
\title{Plot method}
\description{
Plotting method for \code{\link{calib.fit}} and \code{\link{calib}}Provides plotting methods 
}
\section{Methods}{
\describe{

\item{x = "ANY", y = "ANY"}{Objects other than \code{calib.fit} and \code{calib} for which \code{plot} can be used.}

\item{x = "calib", y = "missing"}{\code{calib} object.}

\item{x = "calib.fit", y = "missing"}{\code{calib.fit} object.}
}}
\keyword{methods}

