\name{residuals-methods}
\docType{methods}
\alias{residuals-methods}
\alias{residuals,ANY-method}
\title{Methods for Function residuals in Package calib}
\description{Returns residuals from an object \code{calib.fit}.
}
\section{Methods}{
\describe{

\item{object = "ANY"}{Objects other than \code{calib.fit} object.}

\item{object = "calib.fit"}{\code{calib.fit} object.}
}}
\keyword{methods}
